requirements.txt
alembic==1.13.3
aniso8601==9.0.1
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
distlib==0.3.9
Faker==30.8.1
filelock==3.16.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
honcho==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
marshmallow==3.23.0
packaging==24.1
pipenv==2024.2.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
virtualenv==20.27.0
Werkzeug==3.0.5
